Update neovim config for c/c++
CCLS and neoinclude seem to be sufficient for linting.
To make CCLS work I had to add a local .vimrc that
set values for C_INCLUDE_PATH and CPLUS_INCLUDE_PATH
so that when editting a header file (i.e. something NOT
included in the compile_commands.json), CCLS would know
what to do with it. This feels like a bit of a hack, but
since header files are not compiled objects I am not sure
what else to do. It would obviously be better to not have
any file paths that are not set by the build system.