Add notes on ccls installation
getting ccls language server to play nicely with LLVM for header completion/detection
on macOS seems a little tricky with the default "brew install ccls". This may be
a deficiency in my neovim configuration for ccls/ALE, but I think that it has to do
with how ccls is compiled to be aware of the llvm/include path. Building ccls manually
seems to work just fine though.